Separate logic for quantity timeseries
Prerendering and formatting logic still are for timeseries only
containing quantity values. Yes, timeseries controller provides
also value types other than quantity, and formatting should be
refactored later on to support other value types as well.

However, logic is better separated from profile, trajectory, etc.
which cannot be handled, anyway.
