+++ rest/src/main/java/org/n52/web/ctrl/data/TimeseriesDataController.java
@@ -34,9 +34,11 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.n52.io.Constants;
+import org.n52.io.PreRenderingJob;
 import org.n52.io.handler.DefaultIoFactory;
 import org.n52.io.request.IoParameters;
 import org.n52.io.request.Parameters;
+import org.n52.io.request.RequestStyledParameterSet;
 import org.n52.io.response.dataset.AbstractValue;
 import org.n52.io.response.dataset.Data;
 import org.n52.io.response.dataset.DatasetOutput;
@@ -44,9 +46,12 @@ import org.n52.series.spi.srv.DataService;
 import org.n52.series.spi.srv.ParameterService;
 import org.n52.web.ctrl.UrlSettings;
 import org.n52.web.exception.ResourceNotFoundException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
@@ -59,12 +64,109 @@ import org.springframework.web.servlet.ModelAndView;
     "application/json"
 })
 public class TimeseriesDataController extends DataController {
+    
+    private static final Logger LOGGER = LoggerFactory.getLogger(TimeseriesDataController.class);
+
+    private PreRenderingJob preRenderingTask;
 
     @Autowired
     public TimeseriesDataController(DefaultIoFactory<DatasetOutput<AbstractValue<?>>, AbstractValue<?>> ioFactory,
             ParameterService<DatasetOutput<AbstractValue<?>>> datasetService,
             DataService<Data<AbstractValue<?>>> dataService) {
         super(ioFactory, datasetService, dataService);
+    }
+
+    @RequestMapping(value = "/observations",
+        produces = {
+            Constants.IMAGE_PNG
+        },
+        method = RequestMethod.POST)
+    public void getSeriesCollectionChart(HttpServletRequest request,
+                                         HttpServletResponse response,
+                                         @RequestHeader(value = Parameters.HttpHeader.ACCEPT_LANGUAGE,
+                                             required = false) String locale,
+                                         @RequestBody RequestStyledParameterSet simpleParameters)
+            throws Exception {
+        IoParameters parameters = createParameters(simpleParameters, locale, response);
+        LOGGER.debug("get data collection chart with query: {}", parameters);
+        checkForUnknownDatasetIds(parameters, parameters.getDatasets());
+
+        // final String datasetType = getValueType(parameters);
+        final String valueType = getValueType(parameters, request.getRequestURI());
+        String outputFormat = Constants.IMAGE_PNG;
+        response.setContentType(outputFormat);
+        createIoFactory(valueType).setParameters(parameters)
+                                  .createHandler(outputFormat)
+                                  .writeBinary(response.getOutputStream());
+    }
+
+    @RequestMapping(value = "/{datasetId}/observations",
+        produces = {
+            Constants.IMAGE_PNG
+        },
+        method = RequestMethod.GET)
+    public void getSeriesChart(HttpServletRequest request,
+                               HttpServletResponse response,
+                               @PathVariable String datasetId,
+                               @RequestHeader(value = Parameters.HttpHeader.ACCEPT_LANGUAGE,
+                                   required = false) String locale,
+                               @RequestParam(required = false) MultiValueMap<String, String> query)
+            throws Exception {
+        IoParameters parameters = createParameters(datasetId, query, locale, response);
+        LOGGER.debug("get data collection chart for '{}' with query: {}", datasetId, parameters);
+        checkAgainstTimespanRestriction(parameters.getTimespan());
+        checkForUnknownDatasetId(parameters, datasetId);
+
+        // String handleAsValueTypeFallback = parameters.getAsString(Parameters.HANDLE_AS_VALUE_TYPE);
+        // String valueType = ValueType.extractType(datasetId, handleAsValueTypeFallback);
+        String valueType = getValueType(parameters, request.getRequestURI());
+        String outputFormat = Constants.IMAGE_PNG;
+        response.setContentType(outputFormat);
+        createIoFactory(valueType).setParameters(parameters)
+                                  .createHandler(outputFormat)
+                                  .writeBinary(response.getOutputStream());
+    }
+
+    @RequestMapping(value = "/{datasetId}/images", method = RequestMethod.GET)
+    public ModelAndView getSeriesChartByInterval(@PathVariable String datasetId) {
+        assertPrerenderingIsEnabled();
+        ModelAndView response = new ModelAndView();
+        return response.addObject(preRenderingTask.getPrerenderedImages(datasetId));
+    }
+
+    @RequestMapping(value = "/{datasetId}/{chartQualifier}",
+        produces = {
+            Constants.IMAGE_PNG
+        },
+        method = RequestMethod.GET)
+    public void getSeriesChartByInterval(HttpServletResponse response,
+                                         @PathVariable String datasetId,
+                                         @PathVariable String chartQualifier)
+            throws Exception {
+        assertPrerenderingIsEnabled();
+        assertPrerenderedImageIsAvailable(datasetId, chartQualifier);
+
+        response.setContentType(Constants.IMAGE_PNG);
+        LOGGER.debug("get prerendered chart for '{}' ({})", datasetId, chartQualifier);
+        preRenderingTask.writePrerenderedGraphToOutputStream(datasetId, chartQualifier, response.getOutputStream());
+    }
+
+    @RequestMapping(value = "/{datasetId}/images/{fileName}",
+        produces = {
+            Constants.IMAGE_PNG
+        },
+        method = RequestMethod.GET)
+    public void getSeriesChartByFilename(HttpServletResponse response,
+                                         @PathVariable String datasetId,
+                                         @PathVariable String fileName)
+            throws Exception {
+        assertPrerenderingIsEnabled();
+        assertPrerenderedImageIsAvailable(fileName, null);
+
+        response.setContentType(Constants.IMAGE_PNG);
+        LOGGER.debug("get prerendered chart for '{}'", fileName);
+        preRenderingTask.writePrerenderedGraphToOutputStream(fileName, response.getOutputStream());
+
     }
 
     @Deprecated
@@ -99,4 +201,29 @@ public class TimeseriesDataController extends DataController {
                 : item.getValueType();
     }
 
+    // TODO set preredering config instead of task
+
+    public PreRenderingJob getPreRenderingTask() {
+        return preRenderingTask;
+    }
+
+    public void setPreRenderingTask(PreRenderingJob prerenderingTask) {
+        this.preRenderingTask = prerenderingTask;
+    }
+
+    private void assertPrerenderingIsEnabled() {
+        if (preRenderingTask == null) {
+            throw new ResourceNotFoundException("Diagram prerendering is not enabled.");
+        }
+    }
+
+    private void assertPrerenderedImageIsAvailable(String seriesId, String chartQualifier) {
+        if ( !preRenderingTask.hasPrerenderedImage(seriesId, chartQualifier)) {
+            throw new ResourceNotFoundException("No pre-rendered chart found for datasetId '"
+                    + seriesId
+                    + " (qualifier: "
+                    + chartQualifier
+                    + ")'.");
+        }
+    }
 }
